Don't need to clear local storage, disable some suites in FF

+++ e2e/cypress/support/commands.js
@@ -39,7 +39,9 @@ Cypress.Commands.overwrite("visit", (originalFn, path, options = {}) => {
 
 Cypress.Commands.add("clearSessionStorage", () => {
   cy.window().then((window) => {
-    window.sessionStorage.clear();
+    if (window.sessionStorage) {
+      window.sessionStorage.clear();
+    }
   });
 });
